Remove `DenseArray` restriction for `StridedView`s
This addresses #2.

In particular, while this allows `StridedView`s to be used for non-`DenseArray` types, it does not define a constructor for them.
This reflects the fact that in general, `StridedView`s are only well-defined for `DenseArray`s, but allows users to manually tap into the machinery of `Strided`, in which case it is up to them to ensure correctness.